Fix logic around prefixed branches
The action deletes all branches by default except the protected ones.
As such, there is no reason to have a specific filter to include more
branches into this logic. What we want is a mechanism to *exclude*
branches from this logic in order to artifically protect branches which
are not marked as protected in the github API.